# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.AddOrganizationMemberRequest:
    properties:
      role:
        example: member
        type: string
      user_id:
        type: string
    required:
    - role
    - user_id
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.CreateFunctionRequest:
    properties:
      code:
        type: string
      name:
        example: my_function
        type: string
      runtime:
        example: nodejs
        type: string
    required:
    - code
    - name
    - runtime
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.CreateOrganizationInvitationRequest:
    properties:
      email:
        example: user@example.com
        type: string
      role:
        example: member
        type: string
    required:
    - email
    - role
    type: object
  github_com_ketsuna-org_sovrabase_internal_models_requests.CreateOrganizationRequest:
    properties:
      name:
        example: My Organization
        type: string
    required:
    - name
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.InvokeFunctionRequest:
    properties:
      args:
        additionalProperties: true
        type: object
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateFunctionRequest:
    properties:
      code:
        type: string
      name:
        example: updated_function
        type: string
      runtime:
        example: python
        type: string
    type: object
  github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateOrganizationMemberRequest:
    properties:
      role:
        example: admin
        type: string
    required:
    - role
    type: object
  github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateOrganizationRequest:
    properties:
      name:
        example: Updated Organization
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the Sovrabase API server.
  title: Sovrabase API
  version: "1.0"
paths:
...
  /organization:
    get:
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Organizations
      tags:
      - Organisations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create Organization
      tags:
      - Organisations
  /organization/{id}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete Organization
      tags:
      - Organisations
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Organization update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Update Organization
      tags:
      - Organisations
  /organization/{id}/invitations:
    get:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Organization Invitations
      tags:
      - Organisations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Invitation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.CreateOrganizationInvitationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create Organization Invitation
      tags:
      - Organisations
  /organization/{id}/invitations/{invite_id}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Invitation ID
        in: path
        name: invite_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete Organization Invitation
      tags:
      - Organisations
  /organization/{id}/members:
    get:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Organization Members
      tags:
      - Organisations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Member data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.AddOrganizationMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Add Organization Member
      tags:
      - Organisations
  /organization/{id}/members/{user_id}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete Organization Member
      tags:
      - Organisations
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Member update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateOrganizationMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Update Organization Member
      tags:
      - Organisations
...
  /project/{id}/functions:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: List Functions
      tags:
      - Functions
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Function creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.CreateFunctionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create Function
      tags:
      - Functions
  /project/{id}/functions/{function_id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Function ID
        in: path
        name: function_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete Function
      tags:
      - Functions
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Function ID
        in: path
        name: function_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Function
      tags:
      - Functions
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Function ID
        in: path
        name: function_id
        required: true
        type: string
      - description: Function update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateFunctionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Update Function
      tags:
      - Functions
  /project/{id}/functions/{function_id}/invoke:
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Function ID
        in: path
        name: function_id
        required: true
        type: string
      - description: Function invocation arguments
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.InvokeFunctionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Invoke Function
      tags:
      - Functions
  /project/{id}/functions/{function_id}/logs:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Function ID
        in: path
        name: function_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Execution logs from a function (List of call by usages..)
      tags:
      - Functions
...
swagger: "2.0"