# internal/api/handlers/functions.go
package handlers

import (
	"net/http"

	"github.com/ketsuna-org/sovrabase/internal/models/requests"
)

// ListFunctionsHandler lists all functions in a project
// @Summary List Functions
// @Tags Functions
// @Security Bearer
// @Param id path string true "Project ID"
// @Success 200
// @Router /project/{id}/functions [get]
func ListFunctionsHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement list functions logic
	w.WriteHeader(http.StatusOK)
}

// CreateFunctionHandler creates a new function
// @Summary Create Function
// @Tags Functions
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param request body requests.CreateFunctionRequest true "Function creation data"
// @Success 200
// @Router /project/{id}/functions [post]
func CreateFunctionHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.CreateFunctionRequest
	_ = req
	// TODO: Implement create function logic
	w.WriteHeader(http.StatusOK)
}

// GetFunctionHandler gets a specific function
// @Summary Get Function
// @Tags Functions
// @Security Bearer
// @Param id path string true "Project ID"
// @Param function_id path string true "Function ID"
// @Success 200
// @Router /project/{id}/functions/{function_id} [get]
func GetFunctionHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get function logic
	w.WriteHeader(http.StatusOK)
}

// UpdateFunctionHandler updates a function
// @Summary Update Function
// @Tags Functions
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param function_id path string true "Function ID"
// @Param request body requests.UpdateFunctionRequest true "Function update data"
// @Success 200
// @Router /project/{id}/functions/{function_id} [patch]
func UpdateFunctionHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.UpdateFunctionRequest
	_ = req
	// TODO: Implement update function logic
	w.WriteHeader(http.StatusOK)
}

// DeleteFunctionHandler deletes a function
// @Summary Delete Function
// @Tags Functions
// @Security Bearer
// @Param id path string true "Project ID"
// @Param function_id path string true "Function ID"
// @Success 200
// @Router /project/{id}/functions/{function_id} [delete]
func DeleteFunctionHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement delete function logic
	w.WriteHeader(http.StatusOK)
}

// InvokeFunctionHandler invokes a function
// @Summary Invoke Function
// @Tags Functions
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param function_id path string true "Function ID"
// @Param request body requests.InvokeFunctionRequest true "Function invocation arguments"
// @Success 200
// @Router /project/{id}/functions/{function_id}/invoke [post]
func InvokeFunctionHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.InvokeFunctionRequest
	_ = req
	// TODO: Implement invoke function logic
	w.WriteHeader(http.StatusOK)
}

// GetFunctionLogsHandler gets execution logs for a function
// @Summary Execution logs from a function (List of call by usages..)
// @Tags Functions
// @Security Bearer
// @Param id path string true "Project ID"
// @Param function_id path string true "Function ID"
// @Success 200
// @Router /project/{id}/functions/{function_id}/logs [get]
func GetFunctionLogsHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get function logs logic
	w.WriteHeader(http.StatusOK)
}