# internal/api/handlers/organization.go
package handlers

import (
	"net/http"

	"github.com/ketsuna-org/sovrabase/internal/models/requests"
)

// GetOrganizationsHandler gets all organizations
// @Summary Get Organizations
// @Tags Organisations
// @Security Bearer
// @Success 200
// @Router /organization [get]
func GetOrganizationsHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get organizations logic
	w.WriteHeader(http.StatusOK)
}

// CreateOrganizationHandler creates a new organization
// @Summary Create Organization
// @Tags Organisations
// @Security Bearer
// @Accept json
// @Produce json
// @Param request body requests.CreateOrganizationRequest true "Organization creation data"
// @Success 200
// @Router /organization [post]
func CreateOrganizationHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.CreateOrganizationRequest
	_ = req
	// TODO: Implement create organization logic
	w.WriteHeader(http.StatusOK)
}

// UpdateOrganizationHandler updates an organization
// @Summary Update Organization
// @Tags Organisations
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Organization ID"
// @Param request body requests.UpdateOrganizationRequest true "Organization update data"
// @Success 200
// @Router /organization/{id} [patch]
func UpdateOrganizationHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.UpdateOrganizationRequest
	_ = req
	// TODO: Implement update organization logic
	w.WriteHeader(http.StatusOK)
}

// DeleteOrganizationHandler deletes an organization
// @Summary Delete Organization
// @Tags Organisations
// @Security Bearer
// @Param id path string true "Organization ID"
// @Success 200
// @Router /organization/{id} [delete]
func DeleteOrganizationHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement delete organization logic
	w.WriteHeader(http.StatusOK)
}

// GetOrganizationMembersHandler gets members of an organization
// @Summary Get Organization Members
// @Tags Organisations
// @Security Bearer
// @Param id path string true "Organization ID"
// @Success 200
// @Router /organization/{id}/members [get]
func GetOrganizationMembersHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get members logic
	w.WriteHeader(http.StatusOK)
}

// AddOrganizationMemberHandler adds a member to an organization
// @Summary Add Organization Member
// @Tags Organisations
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Organization ID"
// @Param request body requests.AddOrganizationMemberRequest true "Member data"
// @Success 200
// @Router /organization/{id}/members [post]
func AddOrganizationMemberHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.AddOrganizationMemberRequest
	_ = req
	// TODO: Implement add member logic
	w.WriteHeader(http.StatusOK)
}

// UpdateOrganizationMemberHandler updates an organization member
// @Summary Update Organization Member
// @Tags Organisations
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Organization ID"
// @Param user_id path string true "User ID"
// @Param request body requests.UpdateOrganizationMemberRequest true "Member update data"
// @Success 200
// @Router /organization/{id}/members/{user_id} [patch]
func UpdateOrganizationMemberHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.UpdateOrganizationMemberRequest
	_ = req
	// TODO: Implement update member logic
	w.WriteHeader(http.StatusOK)
}

// DeleteOrganizationMemberHandler removes a member from an organization
// @Summary Delete Organization Member
// @Tags Organisations
// @Security Bearer
// @Param id path string true "Organization ID"
// @Param user_id path string true "User ID"
// @Success 200
// @Router /organization/{id}/members/{user_id} [delete]
func DeleteOrganizationMemberHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement remove member logic
	w.WriteHeader(http.StatusOK)
}

// GetOrganizationInvitationsHandler gets invitations for an organization
// @Summary Get Organization Invitations
// @Tags Organisations
// @Security Bearer
// @Param id path string true "Organization ID"
// @Success 200
// @Router /organization/{id}/invitations [get]
func GetOrganizationInvitationsHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get invitations logic
	w.WriteHeader(http.StatusOK)
}

// CreateOrganizationInvitationHandler creates an invitation
// @Summary Create Organization Invitation
// @Tags Organisations
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Organization ID"
// @Param request body requests.CreateOrganizationInvitationRequest true "Invitation data"
// @Success 200
// @Router /organization/{id}/invitations [post]
func CreateOrganizationInvitationHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.CreateOrganizationInvitationRequest
	_ = req
	// TODO: Implement create invitation logic
	w.WriteHeader(http.StatusOK)
}

// DeleteOrganizationInvitationHandler deletes an invitation
// @Summary Delete Organization Invitation
// @Tags Organisations
// @Security Bearer
// @Param id path string true "Organization ID"
// @Param invite_id path string true "Invitation ID"
// @Success 200
// @Router /organization/{id}/invitations/{invite_id} [delete]
func DeleteOrganizationInvitationHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement delete invitation logic
	w.WriteHeader(http.StatusOK)
}

// GetOrganizationMetricsHandler gets organization metrics
// @Summary Metrix related to an orginasation
// @Tags Organisations
// @Security Bearer
// @Param id path string true "Organization ID"
// @Success 200
// @Router /organization/{id}/metrics [get]
func GetOrganizationMetricsHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get organization metrics logic
	w.WriteHeader(http.StatusOK)
}